Refactor: Rename news search tool file
